initial solution for day 09

+++ src/aoc-2019-csharp/Day02/Day02.cs
@@ -4,11 +4,11 @@ namespace aoc_2019_csharp.Day02;
 
 public static class Day02
 {
-    private static readonly int[] Input = File.ReadAllText("Day02/day02.txt").Split(',').Select(int.Parse).ToArray();
+    private static readonly long[] Input = File.ReadAllText("Day02/day02.txt").Split(',').Select(long.Parse).ToArray();
 
-    public static int Part1() => new IntcodeComputer(Input).Run(12, 2);
+    public static long Part1() => new IntcodeComputer(Input).Run(12, 2);
 
-    public static int Part2()
+    public static long Part2()
     {
         for (var i = 0; i < 100; i++)
         {
