+++ src/aoc-2019-csharp/Day05/Day05.cs
@@ -4,13 +4,13 @@ namespace aoc_2019_csharp.Day05;
 
 public static class Day05
 {
-    private static readonly int[] Input = File.ReadAllText("Day05/day05.txt").Split(',').Select(int.Parse).ToArray();
+    private static readonly long[] Input = File.ReadAllText("Day05/day05.txt").Split(',').Select(long.Parse).ToArray();
 
-    public static int Part1() => Run(1);
+    public static long Part1() => Run(1);
 
-    public static int Part2() => Run(5);
+    public static long Part2() => Run(5);
 
-    private static int Run(int input)
+    private static long Run(long input)
     {
         var inputs = new[] { input };
         var computer = new IntcodeComputer(Input, inputs);
